+++ packages/elements/reactions/src/__tests__/unit/reaction-store/ReactionsStore.test.ts
@@ -1,5 +1,3 @@
-import { shallow, ShallowWrapper } from 'enzyme';
-import * as React from 'react';
 import { ReactionClient } from '../../../client';
 import {
   ari,
@@ -7,12 +5,7 @@ import {
   reaction,
   user,
 } from '../../../client/MockReactionsClient';
-import {
-  Props,
-  ReactionContext,
-  State,
-} from '../../../reaction-store/ReactionContext';
-import { Actions, Context } from '../../../reaction-store/Context';
+import { MemoryReactionsStore } from '../../../reaction-store/ReactionsStore';
 import { ReactionStatus } from '../../../types/ReactionStatus';
 
 describe('ReactionContext', () => {
@@ -34,23 +27,10 @@ describe('ReactionContext', () => {
     ],
   });
 
-  const extractActions = (provider: ShallowWrapper<Props, State>): Actions =>
-    provider.find(Context.Provider).prop('value').actions;
-
-  const renderProvider = () =>
-    shallow(
-      <ReactionContext
-        client={fakeClient}
-        url="http://reactions.atlassian.com"
-      />,
-    );
-
-  let provider: ShallowWrapper<Props, State>;
-  let actions: Actions;
+  let store: MemoryReactionsStore;
 
   beforeEach(() => {
-    provider = renderProvider();
-    actions = extractActions(provider);
+    store = new MemoryReactionsStore(fakeClient);
   });
 
   afterEach(() => {
@@ -60,71 +40,103 @@ describe('ReactionContext', () => {
     (fakeClient.deleteReaction as jest.Mock<any>).mockClear();
   });
 
-  it('should provide actions and state', () => {
-    const contextProvider = provider.find(Context.Provider);
-    expect(contextProvider.prop('value')).toMatchObject({
-      value: {
+  describe('with empty state', () => {
+    beforeEach(() => {
+      store = new MemoryReactionsStore(fakeClient);
+    });
+
+    it('should set initial state', () => {
+      expect(store.getState()).toMatchObject({
         reactions: {},
         flash: {},
-      },
-      actions: {
-        getReactions: expect.any(Function),
-        toggleReaction: expect.any(Function),
-        addReaction: expect.any(Function),
-        getDetailedReaction: expect.any(Function),
-      },
+      });
     });
-  });
 
-  it('should update client', () => {
-    const newClient: ReactionClient = {
-      getReactions: jest.fn(),
-      getDetailedReaction: jest.fn(),
-      addReaction: jest.fn(),
-      deleteReaction: jest.fn(),
-    };
-    (newClient.getReactions as jest.Mock<any>).mockReturnValueOnce(
-      getReactionsResponse,
-    );
-    provider.setProps({ client: newClient });
-    actions.getReactions(containerAri, ari);
+    it('should call client to get reactions', () => {
+      (fakeClient.getReactions as jest.Mock<any>).mockReturnValueOnce(
+        getReactionsResponse,
+      );
 
-    jest.runAllTimers();
-    expect(newClient.getReactions).toHaveBeenCalled();
-    expect(fakeClient.getReactions).not.toHaveBeenCalled();
-  });
+      store.getReactions(containerAri, ari);
 
-  it('should call adaptor to get reactions', () => {
-    (fakeClient.getReactions as jest.Mock<any>).mockReturnValueOnce(
-      getReactionsResponse,
-    );
+      jest.runAllTimers();
 
-    actions.getReactions(containerAri, ari);
+      expect(fakeClient.getReactions).toHaveBeenCalledTimes(1);
 
-    jest.runAllTimers();
-
-    expect(fakeClient.getReactions).toHaveBeenCalledTimes(1);
-
-    return getReactionsResponse.then(() => {
-      expect(provider.state()).toMatchObject({
-        reactions: {
-          [`${containerAri}|${ari}`]: {
-            status: ReactionStatus.ready,
-            reactions: [
-              reaction(':thumbsup:', 3, false),
-              reaction(':fire:', 2, true),
-              reaction(':thumbsdown:', 1, true),
-            ],
+      return getReactionsResponse.then(() => {
+        expect(store.getState()).toMatchObject({
+          reactions: {
+            [`${containerAri}|${ari}`]: {
+              status: ReactionStatus.ready,
+              reactions: [
+                reaction(':thumbsup:', 3, false),
+                reaction(':fire:', 2, true),
+                reaction(':thumbsdown:', 1, true),
+              ],
+            },
           },
-        },
-        flash: {},
+          flash: {},
+        });
+      });
+    });
+
+    it('should notify notify onUpdate', () => {
+      (fakeClient.getReactions as jest.Mock<any>).mockReturnValueOnce(
+        getReactionsResponse,
+      );
+
+      const callback = jest.fn();
+      store.onChange(callback);
+
+      store.getReactions(containerAri, ari);
+
+      jest.runAllTimers();
+
+      return getReactionsResponse.then(() => {
+        // we need to run all timers because onUpdate notification is batched
+        jest.runAllTimers();
+        expect(callback).toHaveBeenCalledTimes(1);
+        expect(callback).toHaveBeenCalledWith({
+          reactions: {
+            [`${containerAri}|${ari}`]: {
+              status: ReactionStatus.ready,
+              reactions: [
+                reaction(':thumbsup:', 3, false),
+                reaction(':fire:', 2, true),
+                reaction(':thumbsdown:', 1, true),
+              ],
+            },
+          },
+          flash: {},
+        });
+      });
+    });
+
+    it('should not notify after removing the callback', () => {
+      (fakeClient.getReactions as jest.Mock<any>).mockReturnValueOnce(
+        getReactionsResponse,
+      );
+
+      const callback = jest.fn();
+      store.onChange(callback);
+
+      store.removeOnChangeListener(callback);
+
+      store.getReactions(containerAri, ari);
+
+      jest.runAllTimers();
+
+      return getReactionsResponse.then(() => {
+        // we need to run all timers because onUpdate notification is batched
+        jest.runAllTimers();
+        expect(callback).not.toHaveBeenCalled();
       });
     });
   });
 
   describe('with state set', () => {
     beforeEach(() => {
-      provider.setState({
+      store = new MemoryReactionsStore(fakeClient, {
         reactions: {
           [`${containerAri}|${ari}`]: {
             reactions: [
@@ -134,6 +146,7 @@ describe('ReactionContext', () => {
             status: ReactionStatus.ready,
           },
         },
+        flash: {},
       });
     });
 
@@ -147,12 +160,12 @@ describe('ReactionContext', () => {
         response,
       );
 
-      actions.getDetailedReaction(containerAri, ari, '1f44d');
+      store.getDetailedReaction(containerAri, ari, '1f44d');
 
       expect(fakeClient.getDetailedReaction).toHaveBeenCalledTimes(1);
 
       return response.then(() => {
-        expect(provider.state()).toMatchObject({
+        expect(store.getState()).toMatchObject({
           reactions: {
             [`${containerAri}|${ari}`]: {
               status: ReactionStatus.ready,
@@ -174,9 +187,9 @@ describe('ReactionContext', () => {
 
       (fakeClient.addReaction as jest.Mock<any>).mockReturnValueOnce(response);
 
-      actions.addReaction(containerAri, ari, '1f44d');
+      store.addReaction(containerAri, ari, '1f44d');
 
-      expect(provider.state()).toMatchObject({
+      expect(store.getState()).toMatchObject({
         reactions: {
           [`${containerAri}|${ari}`]: {
             status: ReactionStatus.ready,
@@ -197,9 +210,9 @@ describe('ReactionContext', () => {
 
       (fakeClient.addReaction as jest.Mock<any>).mockReturnValueOnce(response);
 
-      actions.toggleReaction(containerAri, ari, '1f44d');
+      store.toggleReaction(containerAri, ari, '1f44d');
 
-      expect(provider.state()).toMatchObject({
+      expect(store.getState()).toMatchObject({
         reactions: {
           [`${containerAri}|${ari}`]: {
             status: ReactionStatus.ready,
@@ -216,9 +229,9 @@ describe('ReactionContext', () => {
     });
 
     it('should flash reaction when the user tries to add it again', () => {
-      actions.addReaction(containerAri, ari, '1f44e');
+      store.addReaction(containerAri, ari, '1f44e');
 
-      expect(provider.state()).toMatchObject({
+      expect(store.getState()).toMatchObject({
         reactions: {
           [`${containerAri}|${ari}`]: {
             status: ReactionStatus.ready,
@@ -243,9 +256,9 @@ describe('ReactionContext', () => {
 
       (fakeClient.addReaction as jest.Mock<any>).mockReturnValueOnce(response);
 
-      actions.toggleReaction(containerAri, ari, '1f44e');
+      store.toggleReaction(containerAri, ari, '1f44e');
 
-      expect(provider.state()).toMatchObject({
+      expect(store.getState()).toMatchObject({
         reactions: {
           [`${containerAri}|${ari}`]: {
             status: ReactionStatus.ready,