+++ packages/elements/task-decision/src/__tests__/unit/api/TaskDecisionResourceSpec.ts
@@ -484,18 +484,17 @@ describe('TaskDecisionResource', () => {
         latestState = state;
       };
       resource.subscribe(key1, handler);
+      let toggleStatePromise;
       return waitUntil(() => latestState === 'TODO')
         .then(() => {
-          resource.toggleTask(key1, 'DONE');
+          toggleStatePromise = resource.toggleTask(key1, 'DONE');
           return waitUntil(() => latestState === 'DONE');
         })
         .then(() => {
           expect(latestState).toBe('DONE');
-          // failure will reset state
-          return waitUntil(() => fetchMock.called('set-task'));
-        })
-        .then(() => {
-          expect(latestState).toBe('TODO');
+          return toggleStatePromise.catch(() => {
+            expect(latestState).toBe('TODO');
+          });
         });
     });